requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0

# Document parsing
markitdown[all]>=0.1.4

# Vector storage
qdrant-client>=1.16.2

# PDF processing
pypdf>=3.17.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.9

# Utilities
openai>=1.12.0

